Relocate container folders to component and delete container folder
